+++ src/pages/auth/forgot-password.vue
@@ -1,18 +1,18 @@
 <route lang="yaml">
-name: lost-device
+name: forgot-password
 meta:
   layout: Auth
-  pageTitle: pageTitle.lostDevice
+  pageTitle: pageTitle.forgotPassword
   noAuth: true
 </route>
 
 <template>
   <AuthBox>
     <form @submit.prevent>
-      <AppTitle class="mb-2">{{ t('lostDevice.title') }}</AppTitle>
+      <AppTitle>{{ t('forgotPassword.title') }}</AppTitle>
 
       <template v-if="!isRequestSuccessful">
-        <p class="mb-4">{{ t('lostDevice.description') }}</p>
+        <p class="mb-4">{{ t('forgotPassword.description') }}</p>
 
         <div class="mb-4">
           <AppInput
@@ -28,9 +28,9 @@ meta:
           v-if="errorCode"
           variant="error"
           class="mb-4"
-          :title="t('lostDevice.errorTitle')"
+          :title="t('forgotPassword.errorTitle')"
         >
-          <p>{{ t('lostDevice.errors.' + errorCode) }}</p>
+          <p>{{ t('forgotPassword.errors.' + errorCode) }}</p>
         </AppNotification>
 
         <AppButton
@@ -38,9 +38,9 @@ meta:
           :disabled="validation.$invalid || loading"
           type="submit"
           class="w-full mb-2"
-          @click="submitLostDevice"
+          @click="submitForgotPassword"
         >
-          {{ t('lostDevice.resetDevice') }}
+          {{ t('forgotPassword.resetPassword') }}
         </AppButton>
 
         <div class="text-center">
@@ -52,7 +52,13 @@ meta:
 
       <template v-else>
         <p class="mb-5 rounded bg-primary-10 p-4">
-          {{ t('lostDevice.emailSent') }}
+          <i18n-t keypath="forgotPassword.message">
+            <template #email>
+              <b>{{ email }}</b>
+            </template>
+          </i18n-t>
+          <br />
+          {{ t('forgotPassword.checkInbox') }}
         </p>
 
         <AppButton class="w-full" variant="link" to="/auth/login">{{
@@ -66,6 +72,7 @@ meta:
 <script lang="ts" setup>
 import { ref } from 'vue';
 import { useI18n } from 'vue-i18n';
+import { useRoute } from 'vue-router';
 import useVuelidate from '@vuelidate/core';
 
 import AppInput from '@components/forms/AppInput.vue';
@@ -74,28 +81,19 @@ import AppTitle from '@components/AppTitle.vue';
 import AuthBox from '@components/AuthBox.vue';
 import AppNotification from '@components/AppNotification.vue';
 
-import ResetSecurityFlow from '@utils/api/reset-security-flow.service';
+import ResetSecurityFlowService from '@utils/api/reset-security-flow.service';
 import { isRequestError } from '@utils/api';
 
 import { RESET_SECURITY_FLOW_ERROR_CODE } from '@enums/reset-security-flow-error-code';
 
 const { t } = useI18n();
+const route = useRoute();
 
 const loading = ref(false);
 const errorCode = ref(RESET_SECURITY_FLOW_ERROR_CODE.NONE);
 const isRequestSuccessful = ref(false);
 
-const props = withDefaults(
-  defineProps<{
-    email: string;
-  }>(),
-  {
-    email: '',
-  }
-);
-
-const email = ref(props.email);
-
+const email = ref(route.query.email?.toString() || '');
 const validation = useVuelidate();
 
 const onError = (err: unknown) => {
@@ -114,19 +112,18 @@ const onError = (err: unknown) => {
   throw err;
 };
 
-const submitLostDevice = async () => {
+const submitForgotPassword = async () => {
   loading.value = true;
   errorCode.value = RESET_SECURITY_FLOW_ERROR_CODE.NONE;
   try {
-    await ResetSecurityFlow.resetDeviceBegin(email.value);
-  } catch (error) {
-    onError(error);
-    isRequestSuccessful.value = false;
+    await ResetSecurityFlowService.resetPasswordBegin(email.value);
+    isRequestSuccessful.value = true;
+  } catch (err) {
+    onError(err);
     loading.value = false;
     return;
+  } finally {
+    loading.value = false;
   }
-
-  isRequestSuccessful.value = true;
-  loading.value = false;
 };
 </script>
